restructure home screen
